requirements.txt
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-Pydantic==0.12.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
